feat: Remodelagem do banco de dados e entidades, correções de nomeclaturas, CRUD de projetos e endpoints de select para microregiao e entidade.

+++ src/main/java/br/gov/es/siscap/exception/ProjetoNaoEncontradoException.java
@@ -2,7 +2,7 @@ package br.gov.es.siscap.exception;
 
 public class ProjetoNaoEncontradoException extends RuntimeException{
 
-    public ProjetoNaoEncontradoException(Integer id) {
+    public ProjetoNaoEncontradoException(Long id) {
         super("Não foi encontrado um projeto com o id [" + id + "]!");
     }
 }
